This is a follow up of https://github.com/chainloop-dev/chainloop/pull/2058

This optimization optimizes the query by avoiding a JOIN between workflows and worflow_runs, which is expensive specially in the case when there are no workflows yet (in this case the join just consists of the whole database contents).
The solution is to first query workflows and then workflow runs. Both queries are optimized.

From (note the 2M rows JOIN):
```
Limit  (cost=0.58..15.46 rows=26 width=236) (actual time=767.852..767.854 rows=0 loops=1)
  ->  Nested Loop Semi Join  (cost=0.58..104130.07 rows=181868 width=236) (actual time=767.850..767.851 rows=0 loops=1)
        Join Filter: (workflow_runs.workflow_id = workflows.id)
        ->  Index Scan using workflow_runs_created_at_index on workflow_runs  (cost=0.43..74112.59 rows=2000544 width=236) (actual time=0.301..624.437 rows=2000559 loops=1)
        ->  Materialize  (cost=0.15..9.32 rows=1 width=16) (actual time=0.000..0.000 rows=0 loops=2000559)
              ->  Nested Loop Semi Join  (cost=0.15..9.31 rows=1 width=16) (actual time=0.024..0.025 rows=0 loops=1)
                    ->  Seq Scan on workflows  (cost=0.00..1.14 rows=1 width=32) (actual time=0.024..0.024 rows=0 loops=1)
                          Filter: ((deleted_at IS NULL) AND (organization_id = '4b1ec0ee-e079-4307-b532-e038cb1fa76a'::uuid))
                          Rows Removed by Filter: 11
                    ->  Index Only Scan using organizations_pkey on organizations  (cost=0.15..8.17 rows=1 width=16) (never executed)
                          Index Cond: (id = '4b1ec0ee-e079-4307-b532-e038cb1fa76a'::uuid)
                          Heap Fetches: 0
Planning Time: 1.654 ms
Execution Time: 768.068 ms
```

To:
```
Nested Loop  (cost=0.15..9.41 rows=11 width=159)
  ->  Index Only Scan using organizations_pkey on organizations  (cost=0.15..8.17 rows=1 width=16)
        Index Cond: (id = '18c3f782-4936-4630-ab8d-20b511366699'::uuid)
  ->  Seq Scan on workflows  (cost=0.00..1.14 rows=11 width=159)
        Filter: ((deleted_at IS NULL) AND (organization_id = '18c3f782-4936-4630-ab8d-20b511366699'::uuid))
```

```
Limit  (cost=0.46..1.55 rows=26 width=236) (actual time=0.778..0.803 rows=26 loops=1)
  ->  Index Scan using workflow_runs_created_at_index on workflow_runs  (cost=0.46..84115.34 rows=2000544 width=236) (actual time=0.776..0.798 rows=26 loops=1)
"        Filter: (workflow_id = ANY ('{95f1b597-20e1-498c-b931-fc6d7331860e,da34f3f3-d672-43ff-b332-094f273a016d,289fa1cd-4e57-4961-af98-44e02afc3507,073a52d7-ffb5-4bf1-9648-a998536e0268,c7b38313-2735-4bc1-8202-ffe60a0acb02,28661725-0b34-4656-b69c-eb0d9ffeeccb,95fb665f-a12f-445b-9808-16e62c4cad93,2025d672-1400-48b9-98dd-f4dc9f98784d,2426605b-622d-4ee9-ba3f-67f7eb6b9dc6,a2353400-5d89-4d63-95c3-2fa903f05e3e,f5e645c7-4112-4926-9d4e-3ced5485183d}'::uuid[]))"
Planning Time: 1.937 ms
Execution Time: 1.222 ms
```